# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AVM CARGO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List packages
      parameters:
        - name: archived
          in: query
          schema:
            type: boolean
          description: Filter by archived status
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: search
          in: query
          schema:
            type: string
          description: Search by tracking number, client name, or phone
        - name: userId
          in: query
          schema:
            type: ["number", "null"]
          description: Filter by user ID (admin only)
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"

    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a new package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a single package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /packages/{id}/status:
    post:
      operationId: addPackageStatus
      tags: [packages]
      summary: Add a status update to a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusInput"
      responses:
        "201":
          description: Status added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}/restore:
    post:
      operationId: restorePackage
      tags: [packages]
      summary: Restore a package from archive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Restored package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [client, admin]
        createdAt:
          type: string
          format: date-time
...
    PackageStatus:
      type: string
      enum:
        - created
        - accepted_china
        - departed_china
        - arrived_almaty
        - departed_almaty
        - arrived_city
        - ready_pickup
        - delivered
    PackageInput:
      type: object
      required: [trackingNumber]
      properties:
        trackingNumber:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        deliveryCost:
          type: ["number", "null"]
        userId:
          type: ["number", "null"]
    PackageUpdate:
      type: object
      properties:
        trackingNumber:
          type: string
        description:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        deliveryCost:
          type: ["number", "null"]
        adminComment:
          type: ["string", "null"]
        archived:
          type: boolean
    StatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PackageStatus"
    StatusHistory:
      type: object
      required: [id, packageId, status, changedAt]
      properties:
        id:
          type: number
        packageId:
          type: number
        status:
          $ref: "#/components/schemas/PackageStatus"
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: ["number", "null"]
    Package:
      type: object
      required: [id, trackingNumber, status, archived, userId, createdAt, updatedAt]
      properties:
        id:
          type: number
        trackingNumber:
          type: string
        description:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        deliveryCost:
          type: ["number", "null"]
        status:
          $ref: "#/components/schemas/PackageStatus"
        adminComment:
          type: ["string", "null"]
        archived:
          type: boolean
        userId:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    PackageDetail:
      type: object
      required: [id, trackingNumber, status, archived, userId, createdAt, updatedAt, history]
      properties:
        id:
          type: number
        trackingNumber:
          type: string
        description:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        deliveryCost:
          type: ["number", "null"]
        status:
          $ref: "#/components/schemas/PackageStatus"
        adminComment:
          type: ["string", "null"]
        archived:
          type: boolean
        userId:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        history:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistory"

    # Stats
...