# lib/api-zod/src/generated/types/packageDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PackageStatus } from './packageStatus';
import type { StatusHistory } from './statusHistory';
import type { User } from './user';

export interface PackageDetail {
  id: number;
  trackingNumber: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  status: PackageStatus;
  /** @nullable */
  adminComment?: string | null;
  archived: boolean;
  userId: number;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
  history: StatusHistory[];
}