# lib/api-zod/src/generated/types/packageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PackageInput {
  /** @minLength 1 */
  trackingNumber: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  /** @nullable */
  userId?: number | null;
}