# lib/api-zod/src/generated/types/packageUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PackageUpdate {
  trackingNumber?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  /** @nullable */
  adminComment?: string | null;
  archived?: boolean;
}