# lib/api-zod/src/generated/types/statusHistory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PackageStatus } from './packageStatus';

export interface StatusHistory {
  id: number;
  packageId: number;
  status: PackageStatus;
  changedAt: Date;
  /** @nullable */
  changedBy?: number | null;
}